chore: move drive train results to respective classes
